requirements.txt
django
django-rest-framework
django-crispy-forms
django-filter